+++ CHANGELOG.md
@@ -7,6 +7,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+[Unreleased]
+------------
+
+### Added
+
+- Wrap long freeform expressions
+
+
 [0.9.3] - 2026-07-08
 --------------------
 
@@ -513,6 +521,7 @@ Phase 1: Core Domain.
 - Numeric literals with decimal point support
 - 373 passing tests
 
+[Unreleased]: https://github.com/wisnij/unitary/compare/v0.9.3...HEAD
 [0.9.3]: https://github.com/wisnij/unitary/compare/v0.9.2...v0.9.3
 [0.9.2]: https://github.com/wisnij/unitary/compare/v0.9.1...v0.9.2
 [0.9.1]: https://github.com/wisnij/unitary/compare/v0.9.0...v0.9.1